* Completely rewrite the alpha busspace to hide the implementation from   the drivers. * Remove legacy inx/outx support from chipset and replace with macros   which call busspace. * Rework pci config accesses to route through the pcib device instead of   calling a MD function directly.
With these changes it is possible to cleanly support machines which have
more than one independantly numbered PCI busses. As a bonus, the new
busspace implementation should be measurably faster than the old one.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@65176 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f

+++ sys/dev/ata/ata-all.c
@@ -519,7 +519,7 @@ ata_pci_alloc_resource(device_t dev, device_t child, int type, int *rid,
 {
     struct ata_pci_softc *sc = device_get_softc(dev);
     int masterdev = pci_get_progif(dev) & PCIP_STORAGE_IDE_MASTERDEV;
-    int unit = (int)device_get_ivars(child);
+    int unit = (intptr_t)device_get_ivars(child);
     int myrid;
 
     if (type == SYS_RES_IOPORT) {
